# openapi.yaml
openapi: 3.1.0
info:
  title: 이커머스 플랫폼 API
  description: |
    ## 이커머스 플랫폼 API 명세서

    온라인 상품 판매 플랫폼의 API 문서입니다.

    ### 주요 기능
    - **상품 카탈로그**: 상품 조회, 재고 확인
    - **장바구니**: 상품 담기, 수량 변경, 삭제
    - **주문 및 결제**: 주문 생성, 잔액 기반 결제
    - **선착순 쿠폰**: 쿠폰 발급 (1인 1매), 사용

    ### 특징
    - 재고 동시성 제어 (비관적 락/낙관적 락)
    - 선착순 쿠폰 동시성 제어 (Redis 분산 락)
    - 트랜잭션 무결성 보장
    - 외부 데이터 플랫폼 연동 (비동기)

    ### 기술 스택
    - Spring Boot 3.x
    - MySQL/PostgreSQL
    - Redis
    - Kafka/RabbitMQ
  contact:
    name: E-commerce API Team
    email: support@example.com
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: 로컬 개발 서버
- url: https://api.example.com
  description: 프로덕션 서버
...
paths:
...
  /api/order-items/{orderItemId}/return:
    post:
      tags:
      - 주문 상품
      summary: 주문 상품 반품
      description: 배송 완료된 주문 상품을 반품합니다.
      operationId: returnOrderItem
      parameters:
      - name: orderItemId
        in: path
        description: 주문 상품 ID
        required: true
        schema:
          type: integer
          format: int64
          exclusiveMinimum: 0
      responses:
        "200":
          description: 주문 상품 반품 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderItemResponse"
        "400":
          description: 반품 불가능한 상태 (배송완료 상태가 아님)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 주문 상품을 찾을 수 없음
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/order-items/{orderItemId}/exchange:
    post:
      tags:
      - 주문 상품
      summary: 주문 상품 교환
      description: 배송 완료된 주문 상품을 교환합니다.
      operationId: exchangeOrderItem
      parameters:
      - name: orderItemId
        in: path
        description: 주문 상품 ID
        required: true
        schema:
          type: integer
          format: int64
          exclusiveMinimum: 0
      responses:
        "200":
          description: 주문 상품 교환 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderItemResponse"
        "400":
          description: 교환 불가능한 상태 (배송완료 상태가 아님) 또는 교환 상품 재고 부족
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 주문 상품을 찾을 수 없음
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/order-items/{orderItemId}/cancel:
    post:
      tags:
      - 주문 상품
      summary: 주문 상품 취소
      description: "ORDERED, PREPARING 상태의 주문 상품을 취소합니다."
      operationId: cancelOrderItem
      parameters:
      - name: orderItemId
        in: path
        description: 주문 상품 ID
        required: true
        schema:
          type: integer
          format: int64
          exclusiveMinimum: 0
      responses:
        "200":
          description: 주문 상품 취소 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderItemResponse"
        "400":
          description: 취소 불가능한 상태 (이미 배송중이거나 배송완료)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 주문 상품을 찾을 수 없음
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/order-items/{orderItemId}:
    get:
      tags:
      - 주문 상품
      summary: 주문 상품 상세 조회
      description: 특정 주문 상품의 상세 정보를 조회합니다.
      operationId: getOrderItem
      parameters:
      - name: orderItemId
        in: path
        description: 주문 상품 ID
        required: true
        schema:
          type: integer
          format: int64
          exclusiveMinimum: 0
      responses:
        "200":
          description: 주문 상품 조회 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderItemResponse"
        "404":
          description: 주문 상품을 찾을 수 없음
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/order-items/orders/{orderId}:
    get:
      tags:
      - 주문 상품
      summary: 주문의 상품 목록 조회
      description: 특정 주문에 포함된 모든 상품 목록을 조회합니다.
      operationId: getOrderItems
      parameters:
      - name: orderId
        in: path
        description: 주문 ID
        required: true
        schema:
          type: integer
          format: int64
          exclusiveMinimum: 0
      responses:
        "200":
          description: 주문 상품 목록 조회 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderItemResponse"
        "404":
          description: 주문을 찾을 수 없음
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: 에러 응답
      properties:
        code:
          type: string
          description: 에러 코드
        message:
          type: string
          description: 에러 메시지
...
    OrderItemResponse:
      type: object
      description: 주문 상품 응답
      properties:
        id:
          type: integer
          format: int64
          description: 주문 상품 ID
        orderId:
          type: integer
          format: int64
          description: 주문 ID
        productId:
          type: integer
          format: int64
          description: 상품 ID
        productName:
          type: string
          description: 상품명
        quantity:
          type: integer
          format: int32
          description: 수량
        unitPrice:
          type: integer
          format: int32
          description: 단가
        subtotal:
          type: integer
          format: int32
          description: 소계
        status:
          type: string
          description: 주문 상품 상태
          enum:
          - ORDERED
          - PREPARING
          - SHIPPED
          - DELIVERED
          - CANCELLED
          - RETURNED
          - EXCHANGED
...